+++ src/app/shared/ngx-filter-pipe/ngx-filter.pipe.spec.ts
@@ -56,6 +56,18 @@ describe('Pipe: FilterPipe', () => {
       .toEqual([{ value: 'abc', nested: { number: 2 } }]);
   });
 
+  it('filters array of objects with nested objects', () => {
+    const array = [
+      { nested: { number: 1 } },
+      { nested: { number: 2 } },
+      { nested: null }
+    ];
+    const filter = { nested: { number: 2 } };
+    const result = [{ nested: { number: 2 } }];
+
+    expect(pipe.transform(array, filter)).toEqual(result);
+  });
+
   it('filters array of objects with nested objects (not every object has nested object)', () => {
     expect(pipe.transform([
       { value: 'a' },
@@ -286,6 +298,15 @@ describe('Pipe: FilterPipe', () => {
       { name: 'asd' },
       { name: 'aad' }
     ]);
+  });
+
+  it('should filter ', () => {
+    const objects = [
+      { name: 'asd' },
+      { firstName: 'asd' }
+    ];
+
+    expect(pipe.transform(objects , { $or: [{ name: 'asd' }, { firstName: 'asd' }] })).toEqual(objects);
   });
 
   // describe('search in string', () => {